# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VENDORA API",
        "description": "API documentation for VENDORA.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application base URL"
        }
    ],
    "paths": {
...
        "/api/ledger": {
            "get": {
                "tags": [
                    "Ledger"
                ],
                "summary": "List ledger entries",
                "operationId": "5def9712507f8cca1a1a3f2e40c56d1c",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ledger entries list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ledger"
                ],
                "summary": "Create a manual ledger entry",
                "operationId": "5d206e6e425991517ade9623c4576d76",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "description"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "stock_in"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 5000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Purchased new stock"
                                    },
                                    "reference": {
                                        "type": "string",
                                        "example": "PO-001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ledger entry created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ledger/summary": {
            "get": {
                "tags": [
                    "Ledger"
                ],
                "summary": "Ledger summary totals",
                "operationId": "315c10f5960309bf87536b805f5e53e2",
                "responses": {
                    "200": {
                        "description": "Ledger summary"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}